requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.3
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.5.0
tzdata==2025.1
Werkzeug==3.1.3
# Deep Learning
tensorflow==2.15.0
torch==2.2.0
transformers==4.37.2
# Computer Vision
opencv-python==4.9.0.80
pillow==10.2.0
# NLP
nltk==3.8.1
spacy==3.7.2
# Graph Processing
networkx==3.2.1
# Time Series
prophet==1.1.5
# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
